+++ CHANGELOG.md
@@ -6,6 +6,7 @@ All notable changes to [table-js](https://github.com/bpmn-io/table-js) are docum
 
 ___Note:__ Yet to be released changes appear here._
 
+* `FIX`: scope drag and drop listeners to the table container instead of `document` ([#106](https://github.com/bpmn-io/table-js/pull/106))
 * `DEPS`: update to `min-dom@5.3.0`
 
 ## 9.4.0